Split the test monolith into individual tests.
